test(maven): uat for maven build + refactor
