+++ .github/workflows/CI.yml
@@ -44,8 +44,12 @@ jobs:
         run: |
           pip install -r test-requirements.txt
           coverage run --source=gdk -m pytest -v -s tests integration_tests && coverage xml --fail-under=90
+      - name: Set up JDK 1.8
+        uses: actions/setup-java@v1
+        with:
+          java-version: 1.8
       - name: Run UATs 
-        run:
+        run: | 
           coverage run --source=gdk -m pytest -v -s uat
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v2
