+++ uat/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+@pytest.fixture()
+def change_test_dir(tmpdir, monkeypatch):
+    monkeypatch.chdir(tmpdir)
+    return tmpdir
